Include "icons" directory.
Implemented Internal delays in control_generic_class.py and refactored "evaluated_time_delay". Notice that now the class recognizes if there is a time delay or a internal delay.

Function get_plots not longer support a list of transfer functions as this feature will be implemented directly through the GUI.

Script "delays" was deleted from the repository as it was only a scratch. then it was remplaced by "ControlPython_GUI.py".
The feedback argument is not longer evaluated in the main ( ControlPython_set_up.py ) as this process was already evaluated in control_generic_class.py.